shared/python-common/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-common"
version = "0.1.0"
description = "Shared package for enterprise AI platform services"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.5",
  "httpx>=0.27.2,<1",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.6.1",
  "PyJWT[crypto]>=2.8.0",
  "psycopg[binary]>=3.2.3",
  "psycopg-pool>=3.2.0",
  "opentelemetry-api>=1.25.0",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-instrumentation-fastapi>=0.46b0",
  "opentelemetry-instrumentation-httpx>=0.46b0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",
  "slowapi>=0.1.9",
  "sse-starlette>=2.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
